+++ Assets/Resources/Player/Player.prefab
@@ -168,7 +168,7 @@ CircleCollider2D:
   m_IsTrigger: 1
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: -0.6, y: -0.25}
+  m_Offset: {x: 0, y: -0.5833}
   serializedVersion: 2
   m_Radius: 0.3
 --- !u!1001 &1554514119164177969
