+++ Assets/Scripts/UI/AxisController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 037d93bb328114d4da403d81dcbcea31
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
