+++ Assets/Scripts/UI/AxisTabController.cs
@@ -2,33 +2,18 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-
-public class AxisTabController : Controller{
-    [SerializeField]
-    string axisName=null;//GetAxisRawに渡す引数
-    private void Awake() {
-        switch (axisName) {
-            case "VerticalDigital":
-                type = ControllerManagers.ControllerType.VERTICAL_TAB;
-                break;
-            case "HorizontalDigital":
-                type = ControllerManagers.ControllerType.HORIZONTAL_TAB;
-                break;
-        }
-    }
+public class AxisTabController : AxisController{
     DS.UI.Tab tab;
     private void Start() {
         tab = GetComponent<DS.UI.Tab>();
-        if (axisName == null) {
-            Debug.LogError(gameObject.name + "のAxisTabControllerで、axisNameが未設定です");
+        if (tab == null) {
+            Debug.LogError("AxisTabControllerがアタッチされた"+gameObject.name+"にTabが見つかりません");
         }
     }
-    private void OnDisable() {
-        ChangeControllToPrevious();
-    }
-    override protected IEnumerator Controll() {
+    
+    protected override IEnumerator Controll() {
         while (true) {
-            var axis = Input.GetAxisRaw(axisName);
+            var axis = Input.GetAxisRaw(base.axisName);
             if (axis > 0.2f) {
                 tab.PreviousTab();
                 yield return new WaitForSecondsRealtime(0.2f);
