+++ Assets/Scripts/UI/AxisTabController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 037d93bb328114d4da403d81dcbcea31
+guid: e003b2069105f11438929ff1fdee0a32
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
